use more descriptive link name
Addressing a review comment https://github.com/sphinx-contrib/emojicodes/pull/12#discussion_r396013077

"In the new version of the documentation, I'm including the README inside
the index.rst. That's why I've changed the wording. When the link is
rendered on GitHub or PyPI, the wording "the project's documentation page"
makes sense, but when it's rendered as part of the docs, it reads strange.

I like more descriptive links, but in this case I compromised on "here"
for the above reasons. Nevertheless, I'll change it to "the project's
documentation page" and we can improve it later if we can think
of something better."